// src/app/(app)/(tabs)/home/types/payment.types.ts
export interface PaymentStatusUpdate {
  paymentResponse: {
    txn_id: string;
    amount: number;
    order_id: string;
    txn_detail: {
      status: 'CHARGED' | 'FAILED';
    };
    payment_gateway_response?: {
      resp_message: string;
    };
  };
}